traverser: use int64_t to track scheduled points
Problem: We currently use int to traverse the
scheduled time points that are interatively returned
from planner_multi_avail_time_*. This can lead
to a 32bit integer overflow bug.

Use an int64_t variable instead in traversing
the scheduled time points.

+++ resource/traversers/dfu.cpp
@@ -34,7 +34,7 @@ int dfu_traverser_t::schedule (Jobspec::Jobspec &jobspec,
                                vtx_t root,
                                std::unordered_map<std::string, int64_t> &dfv)
 {
-    int t = 0;
+    int64_t t = 0;
     int rc = -1;
     size_t len = 0;
     std::vector<uint64_t> agg;
